pyproject.toml
[project]
name = "claw-review"
version = "0.1.0"
description = "AI-powered PR triage using multi-model consensus via OpenRouter"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "click>=8.1",
    "rich>=13.7",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "PyJWT>=2.9",
    "cryptography>=44.0",
]

[project.scripts]
claw-review = "claw_review.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "mypy>=1.8",
    "ruff>=0.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
